UI: Range chart widget config improvements.
